apps/api/requirements-api.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
Jinja2==3.1.4
sentence-transformers==3.2.1
faiss-cpu==1.8.0
transformers==4.46.2
#torch==2.5.1
# scikit-learn==1.5.2
# scipy==1.14.1
requests==2.32.3
Markdown==3.7
numpy<2
accelerate>=0.26.0